Add files via upload
In this project, the approximate energy consumption of two different electric vehicle models is found based on 
the input data given. To achieve this purpose, we have computed the coefficients of different degree polynomials manually 
using the sum of least squares method and verified them using inbuilt functions in Matlab. Calculations and observations 
are done using Matlab software. Based on the observations the ‘best fit’ curve was concluded for different electric vehicles 
models. The range or maximum distance that can be travelled is also computed at different speeds and a comparison is 
made for both the models.